pyproject.toml
[tool.poetry]
name = "volatility-estimator"
version = "0.1.0"
description = "Simple library for historical volatility estimation."
authors = ["Ben Dowling <ben.dowling@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
ipykernel = "^6.29.5"
matplotlib = "^3.9.1"
seaborn = "^0.13.2"
pyarrow = "^17.0.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
watchdog = "^4.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.4"
pytest = "^8.3.1"

[tool.poetry.group.report.dependencies]
ipykernel = "^6.29.5"
nbformat = "^5.10.4"
nbclient = "^0.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
# E - pycodestyle error
# F - pyflakes
# I - isort
# N - pep8-naming
# W - pycodestyle warn
# UP - pyupgrade
# TODO: D - pydocstyle
select = ["E", "F", "I", "N", "W", "UP"]
